training and testing data

+++ src/test_split.py
@@ -18,17 +18,16 @@ import shutil
 #             shutil.move(i, move_to)
 #         except:
 #             pass
-dir = '/Users/chenmo/Files/PythonProjects/cs175_project/hand-text-editor/src/edgedata/training/'
+
+dir = 'edgedata/training/'
 training_set = os.listdir(dir)
-
-print(training_set)
 for letter in training_set[1:]:
+    if ".zip" not in letter:
+        file_list = os.listdir(dir + str(letter))
+        data_size = len(file_list)
+        test_list = random.sample(file_list, k=int(data_size/2))
 
-    file_list = os.listdir(dir + str(letter))
-    data_size = len(file_list)
-    test_list = random.sample(file_list, k=int(data_size/2))
+        to_remove = [dir + letter + '/' + x for x in test_list]
 
-    to_remove = [dir + letter + '/' + x for x in test_list]
-
-    for f in to_remove:
-        os.remove(f)
+        for f in to_remove:
+            os.remove(f)
